# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/custom-detectors/{id}": {
      "get": {
        "operationId": "CustomDetectorsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom detector UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomDetectorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get custom detector by ID",
        "tags": [
          "Custom Detectors"
        ]
      },
      "patch": {
        "operationId": "CustomDetectorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom detector UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomDetectorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomDetectorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update custom detector",
        "tags": [
          "Custom Detectors"
        ]
      },
      "delete": {
        "operationId": "CustomDetectorsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Custom detector UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "summary": "Delete custom detector",
        "tags": [
          "Custom Detectors"
        ]
      }
    },
...
    "/custom-detectors/{detectorId}/test-scenarios": {
      "get": {
        "operationId": "CustomDetectorTestsController_list",
        "parameters": [
          {
            "name": "detectorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CustomDetectorTests"
        ]
      },
      "post": {
        "operationId": "CustomDetectorTestsController_create",
        "parameters": [
          {
            "name": "detectorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CustomDetectorTests"
        ]
      }
    },
    "/custom-detectors/{detectorId}/test-scenarios/{scenarioId}": {
      "delete": {
        "operationId": "CustomDetectorTestsController_delete",
        "parameters": [
          {
            "name": "detectorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scenarioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "CustomDetectorTests"
        ]
      }
    },
    "/custom-detectors/{detectorId}/test-scenarios/run": {
      "post": {
        "operationId": "CustomDetectorTestsController_run",
        "parameters": [
          {
            "name": "detectorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "triggeredBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "CustomDetectorTests"
        ]
      }
    },
...
  },
  "info": {
    "title": "Classifyre API",
    "description": "Metadata ingestion and detection API for unstructured data sources. Supports WordPress, Slack, Sitemap, S3-Compatible Storage, Azure Blob Storage, Google Cloud Storage, PostgreSQL, MySQL, MSSQL, Oracle, Hive, Databricks, Snowflake, MongoDB, PowerBI, Tableau, Confluence, Jira, and Service Desk sources. Built-in detectors for secrets, PII, toxic content, NSFW images, broken links, and security threats.",
    "version": "1.0.0",
    "contact": {
      "name": "Classifyre Team",
      "url": "https://github.com/unstructured/classifyre",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CustomDetectorTrainingRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "customDetectorId": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "strategy": {
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "durationMs": {
            "type": "number"
          },
          "trainedExamples": {
            "type": "number"
          },
          "positiveExamples": {
            "type": "number"
          },
          "negativeExamples": {
            "type": "number"
          },
          "metrics": {
            "type": "object",
            "additionalProperties": true
          },
          "modelArtifactPath": {
            "type": "string"
          },
          "configHash": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "customDetectorId",
          "status",
          "startedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomDetectorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "RULESET",
              "CLASSIFIER",
              "ENTITY"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "version": {
            "type": "number"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          },
          "lastTrainedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastTrainingSummary": {
            "type": "object",
            "additionalProperties": true
          },
          "latestTrainingRun": {
            "$ref": "#/components/schemas/CustomDetectorTrainingRunDto"
          },
          "findingsCount": {
            "type": "number"
          },
          "sourcesUsingCount": {
            "type": "number",
            "description": "Number of sources currently selecting this detector"
          },
          "sourcesWithFindingsCount": {
            "type": "number",
            "description": "Number of distinct sources where this detector produced findings"
          },
          "recentSourceNames": {
            "description": "Recent source names using this detector",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourcesUsing": {
            "type": "array",
            "description": "Sources using this detector with their id and name",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "method",
          "isActive",
          "version",
          "config",
          "findingsCount",
          "sourcesUsingCount",
          "sourcesWithFindingsCount",
          "recentSourceNames",
          "sourcesUsing",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateCustomDetectorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-friendly detector name",
            "example": "DACH Contract Risk Terms"
          },
          "key": {
            "type": "string",
            "description": "Stable key override. If omitted, generated from name.",
            "example": "cust_dach_contract_risk"
          },
          "description": {
            "type": "string",
            "description": "Optional detector description",
            "example": "Detect legal/compliance risk language in contracts"
          },
          "method": {
            "type": "string",
            "enum": [
              "RULESET",
              "CLASSIFIER",
              "ENTITY"
            ],
            "description": "Execution method for this custom detector",
            "example": "CLASSIFIER"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether this detector can be selected in sources",
            "default": true
          },
          "config": {
            "type": "object",
            "description": "Custom detector config payload",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}